requirements.txt
# Google Core
google-adk>=1.0.0
google-genai>=1.0.0
google-auth>=2.0.0

# Browser & Automation
playwright>=1.40.0

# Streamlit Dashboard
streamlit>=1.30.0

# Environment management
python-dotenv>=1.0.0
Pillow>=10.0.0
opencv-python-headless  # [NEW] سرعة معالجة الصور

# Configs & Validation
pyyaml>=6.0
pydantic>=2.0.0         # [NEW] تأمين الإعدادات

# Testing
pytest>=7.0.0           # [NEW] لتشغيل مجلد tests/
pytest-asyncio>=0.21.0  # [NEW] لدعم اختبارات الـ Async